% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_expectation.R
\name{compute_expectation}
\alias{compute_expectation}
\alias{compute_expectation.matrix}
\title{Compute an expectation.}
\usage{
compute_expectation(x, ...)

\method{compute_expectation}{matrix}(
  draws,
  lw,
  expectation_fun,
  transformation_fun = NULL,
  log_expectation_fun = FALSE,
  ...
)
}
\description{
Compute an expectation.
}
