% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrain_draws.R
\name{constrain_draws}
\alias{constrain_draws}
\title{Constrain all draws from a fitted model}
\usage{
constrain_draws(x, ...)
}
\arguments{
\item{x}{model fit object}

\item{...}{arguments passed to methods}
}
\value{
constrained draws
}
\description{
Constrain all draws from a fitted model
}
