% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_iwmm_model.R
\name{example_iwmm_model}
\alias{example_iwmm_model}
\title{Example Stan model for importance weighted moment matching}
\usage{
example_iwmm_model(model = "normal_model")
}
\arguments{
\item{model}{Character specifying which model code to
return. Currently "normal_model" is implemented.}
}
\value{
List containing model code and corresponding data.
}
\description{
Provides example models (with data) that are ready for use with
IWMM.
}
