% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectation_moment_match.R
\name{expectation_moment_match}
\alias{expectation_moment_match}
\title{Generic importance weighted moment matching algorithm for a specific
expectation.}
\usage{
expectation_moment_match(draws, ...)
}
\arguments{
\item{draws}{A matrix of draws.}

\item{...}{Further arguments passed.}
}
\description{
Generic importance weighted moment matching algorithm for a specific
expectation.
}
