% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_prob_draws.R
\name{log_prob_draws}
\alias{log_prob_draws}
\title{Return log probability of posterior}
\usage{
log_prob_draws(fit, ...)
}
\arguments{
\item{fit}{model fit object}

\item{...}{arguments passed to methods}
}
\value{
TODO
}
\description{
Return log probability of posterior
}
