% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfit_functions.R
\name{moment_match.stanfit}
\alias{moment_match.stanfit}
\title{Generic importance weighted moment matching algorithm for \code{stanfit} objects.
See additional arguments from \code{moment_match.matrix}}
\usage{
\method{moment_match}{stanfit}(x, log_prob_target_fun = NULL, log_ratio_fun = NULL, ...)
}
\arguments{
\item{x}{A fitted \code{stanfit} object.}

\item{log_prob_target_fun}{Log density of the target.
The function takes argument \code{draws}.}

\item{log_ratio_fun}{Log of the density ratio (target/proposal).
The function takes argument \code{draws}.}

\item{...}{Further arguments passed to \code{moment_match.matrix}.}
}
\value{
Returns a list with 3 elements: transformed draws, updated
importance weights, and the pareto k diagnostic value.
}
\description{
Generic importance weighted moment matching algorithm for \code{stanfit} objects.
See additional arguments from \code{moment_match.matrix}
}
