% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment_match.R
\name{moment_match_modelfit}
\alias{moment_match_modelfit}
\title{Generic importance weighted moment matching algorithm for matrices.}
\usage{
moment_match_modelfit(
  x,
  post_draws,
  unconstrain_pars,
  log_prob_prop_draws_fun,
  log_prob_target_draws_fun = NULL,
  log_ratio_draws_fun = NULL,
  k_threshold = 0.5,
  cov_transform = TRUE,
  ...
)
}
\arguments{
\item{x}{A fitted model object.}

\item{post_draws}{A function the takes \code{x} as the first argument and returns
a matrix of posterior draws of the model parameters (\code{pars}).}

\item{unconstrain_pars}{A function that takes arguments \code{x}, and \code{pars} and
returns posterior draws on the unconstrained space based on the posterior
draws on the constrained space passed via \code{pars}.}

\item{log_prob_prop_draws_fun}{Log density of the proposal.
The function takes argument \code{draws}.}

\item{log_prob_target_draws_fun}{Log density of the target.
The function takes argument \code{draws}.}

\item{log_ratio_draws_fun}{Log of the density ratio (target/proposal).
The function takes argument \code{draws}.}

\item{k_threshold}{Threshold value for Pareto k values above which the moment
matching algorithm is used. The default value is 0.5.}

\item{cov_transform}{Logical; Indicates whether to match the covariance of
the samples or not. If \code{FALSE}, only the mean and marginal variances are
matched. Default is \code{TRUE}.}

\item{...}{Further arguments passed to \code{log_prob_prop_draws_fun},
\code{log_prob_target_draws_fun} and \code{log_ratio_draws_fun}.}
}
\value{
Returns a list with 3 elements: transformed draws, updated
importance weights, and the pareto k diagnostic value.
}
\description{
Generic importance weighted moment matching algorithm for matrices.
}
