% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{shift}
\alias{shift}
\title{Shift a matrix of draws to their weighted mean.}
\usage{
shift(draws, lw)
}
\arguments{
\item{draws}{A matrix of draws.}

\item{lw}{A vector representing the log-weight of each draw.}
}
\value{
List with the shift that was performed, and the new draws matrix.
}
\description{
Shift a matrix of draws to their weighted mean.
}
