% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{shift_and_cov}
\alias{shift_and_cov}
\title{Shift a matrix of draws to their weighted mean and scale the covariance
to match the weighted covariance.}
\usage{
shift_and_cov(draws, lw)
}
\arguments{
\item{draws}{A matrix of draws.}

\item{lw}{A vector representing the log-weight of each draw.}
}
\value{
List with the shift and mapping that were performed, and the new
draws matrix.
}
\description{
Shift a matrix of draws to their weighted mean and scale the covariance
to match the weighted covariance.
}
