% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{shift_and_scale}
\alias{shift_and_scale}
\title{Shift a matrix of draws to their weighted mean and scale the marginal
variances to match the weighted marginal variances.}
\usage{
shift_and_scale(draws, lw)
}
\arguments{
\item{draws}{A matrix of draws.}

\item{lw}{A vector representing the log-weight of each draw.}
}
\value{
List with the shift and scaling that were performed, and the new
draws matrix.
}
\description{
Shift a matrix of draws to their weighted mean and scale the marginal
variances to match the weighted marginal variances.
}
