% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unconstrain_draws.R
\name{unconstrain_draws}
\alias{unconstrain_draws}
\title{Unconstrain all draws from a fitted model}
\usage{
unconstrain_draws(x, ...)
}
\arguments{
\item{x}{model fit object}

\item{...}{arguments passed to methods}
}
\value{
unconstrained draws
}
\description{
Unconstrain all draws from a fitted model
}
