% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment_match.R
\name{moment_match}
\alias{moment_match}
\title{Generic importance weighted moment matching algorithm.}
\usage{
moment_match(draws, ...)
}
\arguments{
\item{draws}{A matrix of draws.}

\item{...}{Further arguments passed.}
}
\description{
Matches a matrix of draws to its importance weighted moments, but
does not compute any expectation.
}
